[FEAT] Q&A 카테고리 ID를 Long에서 int로 변경 및 관련 코드 수정

+++ src/main/java/com/example/kbuddy_backend/qna/dto/request/QnaSaveRequest.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public record QnaSaveRequest(
         @NotNull(message = "카테고리 ID는 필수입니다")
-        Long categoryId,
+        int categoryId,
         
         @NotBlank(message = "제목은 필수입니다")
         @Size(max = 100, message = "제목은 최대 100자까지 입력 가능합니다")
@@ -18,7 +18,7 @@ public record QnaSaveRequest(
         
         List<String> hashtags) {
     
-    public static QnaSaveRequest of(String title, String description, List<String> hashtags, Long categoryId) {
+    public static QnaSaveRequest of(String title, String description, List<String> hashtags, int categoryId) {
         return new QnaSaveRequest(categoryId, title, description, hashtags);
     }
 }
